fix error

+++ test/legacy_test/test_lu_solve_op.py
@@ -31,54 +31,54 @@ class TestLUAPIError(unittest.TestCase):
         with paddle.base.dygraph.guard():
             # The size of b shoule gather than 2.
             def test_b_size():
-                b = paddle.randn([3], dtype='float32')
-                lu = paddle.randn([3, 3], dtype='float32')
-                pivots = paddle.randn([3], dtype='int32')
+                b = paddle.randn([3])
+                lu = paddle.randn([3, 3])
+                pivots = paddle.randn([3])
                 paddle.linalg.lu_solve(b, lu, pivots)
 
             self.assertRaises(ValueError, test_b_size)
 
             # The size of lu shoule gather than 2.
             def test_lu_size():
-                b = paddle.randn([3, 1], dtype='float32')
-                lu = paddle.randn([3], dtype='float32')
-                pivots = paddle.randn([3], dtype='int32')
+                b = paddle.randn([3, 1])
+                lu = paddle.randn([3])
+                pivots = paddle.randn([3])
                 paddle.linalg.lu_solve(b, lu, pivots)
 
             self.assertRaises(ValueError, test_lu_size)
 
             # The size of pivots shoule gather than 2.
             def test_pivots_size():
-                b = paddle.randn([3, 1], dtype='float32')
-                lu = paddle.randn([3, 3], dtype='float32')
-                pivots = paddle.randn([0], dtype='int32')
+                b = paddle.randn([3, 1])
+                lu = paddle.randn([3, 3])
+                pivots = paddle.randn([0])
                 paddle.linalg.lu_solve(b, lu, pivots)
 
             self.assertRaises(ValueError, test_pivots_size)
 
             # b.shape[-1] shoule equal to lu.shape[-2].
             def test_b_lu_shape():
-                b = paddle.randn([3, 1], dtype='float32')
-                lu = paddle.randn([3, 3], dtype='float32')
-                pivots = paddle.randn([3], dtype='int32')
+                b = paddle.randn([3, 1])
+                lu = paddle.randn([3, 3])
+                pivots = paddle.randn([3])
                 paddle.linalg.lu_solve(b, lu, pivots)
 
             self.assertRaises(ValueError, test_b_lu_shape)
 
             # b.shape[-1] shoule equal to pivots.shape[-1].
             def test_b_pivots_shape():
-                b = paddle.randn([1, 3], dtype='float32')
-                lu = paddle.randn([3, 3], dtype='float32')
-                pivots = paddle.randn([2], dtype='int32')
+                b = paddle.randn([1, 3])
+                lu = paddle.randn([3, 3])
+                pivots = paddle.randn([2])
                 paddle.linalg.lu_solve(b, lu, pivots)
 
             self.assertRaises(ValueError, test_b_pivots_shape)
 
             # lu.shape[-2] shoule equal to lu.shape[-1].
             def test_lu_shape():
-                b = paddle.randn([1, 3], dtype='float32')
-                lu = paddle.randn([3, 2], dtype='float32')
-                pivots = paddle.randn([3], dtype='int32')
+                b = paddle.randn([1, 3])
+                lu = paddle.randn([3, 2])
+                pivots = paddle.randn([3])
                 paddle.linalg.lu_solve(b, lu, pivots)
 
             self.assertRaises(ValueError, test_lu_shape)
